feat: refactor all sponge-based hash functions
