+++ miden-crypto/src/hash/algebraic_sponge/rescue/mod.rs
@@ -1,12 +1,12 @@
-use core::ops::Range;
-
-use super::{CubeExtension, ElementHasher, Felt, FieldElement, Hasher, StarkField, ZERO};
+#[cfg(test)]
+use super::{ALPHA, Hasher, INV_ALPHA};
+use super::{CubeExtension, Felt, FieldElement, STATE_WIDTH, StarkField, ZERO};
 
 mod arch;
 pub use arch::optimized::{add_constants_and_apply_inv_sbox, add_constants_and_apply_sbox};
 
 mod mds;
-use mds::{MDS, apply_mds};
+use mds::apply_mds;
 
 mod rpo;
 pub use rpo::Rpo256;
@@ -20,41 +20,9 @@ mod tests;
 // CONSTANTS
 // ================================================================================================
 
-/// The number of rounds is set to 7. For the RPO hash functions all rounds are uniform. For the
+/// The number of rounds is set to 7. For the RPO hash function all rounds are uniform. For the
 /// RPX hash function, there are 3 different types of rounds.
 const NUM_ROUNDS: usize = 7;
-
-/// Sponge state is set to 12 field elements or 96 bytes; 8 elements are reserved for rate and
-/// the remaining 4 elements are reserved for capacity.
-const STATE_WIDTH: usize = 12;
-
-/// The rate portion of the state is located in elements 4 through 11.
-const RATE_RANGE: Range<usize> = 4..12;
-const RATE_WIDTH: usize = RATE_RANGE.end - RATE_RANGE.start;
-
-const INPUT1_RANGE: Range<usize> = 4..8;
-const INPUT2_RANGE: Range<usize> = 8..12;
-
-/// The capacity portion of the state is located in elements 0, 1, 2, and 3.
-const CAPACITY_RANGE: Range<usize> = 0..4;
-
-/// The output of the hash function is a digest which consists of 4 field elements or 32 bytes.
-///
-/// The digest is returned from state elements 4, 5, 6, and 7 (the first four elements of the
-/// rate portion).
-const DIGEST_RANGE: Range<usize> = 4..8;
-
-/// The number of byte chunks defining a field element when hashing a sequence of bytes
-const BINARY_CHUNK_SIZE: usize = 7;
-
-/// S-Box and Inverse S-Box powers;
-///
-/// The constants are defined for tests only because the exponentiations in the code are unrolled
-/// for efficiency reasons.
-#[cfg(test)]
-const ALPHA: u64 = 7;
-#[cfg(test)]
-const INV_ALPHA: u64 = 10540996611094048183;
 
 // SBOX FUNCTION
 // ================================================================================================
